+++ src/pages/HomePage.js
@@ -1,92 +1,75 @@
 // src/pages/HomePage.js
-import React from 'react';
+
 import Map from '../components/Map';
+import React, { useEffect, useState } from "react";
+import * as XLSX from "xlsx";
 
 const HomePage = () => {
+  const [visitedPlaces, setVisitedPlaces] = useState([]);
+  const [plannedPlaces, setPlannedPlaces] = useState([]);
 
+  // Function to load and parse the Excel file
+  const fetchExcelData = async () => {
+    try {
+      const response = await fetch("itravel/data.xlsx"); // Path to your Excel file
+      const arrayBuffer = await response.arrayBuffer();
+      const workbook = XLSX.read(arrayBuffer, { type: "array" });
 
-  // ------------------------------------------------------------------------------------------------------------- 🚩 VISITED PLACES BELOW
+      // Fetch the data from the "Visited" sheet
+      const visitedSheetName = "Visited";
+      const visitedSheet = workbook.Sheets[visitedSheetName];
+      const visitedData = XLSX.utils.sheet_to_json(visitedSheet);
 
-  const visitedPlaces = [
-    { name: 'Jangtaesan Mountain, Daejeon', coords: [36.219296, 127.3394828], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEgZ0QPcWXw2y6fYK9zH1uDbg-c_8ljZmnyimbGhGEgC28LB2YhvvRo8zdzaCgwfhM6WXEqfF5RZXLQOluPVeuz3245wVEXRJB_5wq47-NPdyp5gszt14kC_OdGYZBcBtDOKzG2G39oFq9IrQzPgwZeFXr4-ACbCBp1y5ypSRUKKMo-BiWbyZi8rtUUPfn7q/s856/jangtaesan.png' },
-    { name: 'Sky Road, Daejeon', coords: [36.3271134, 127.4289631], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEi7Cnm__rfxdFFhbuYl42ExWaS_lGzIqsmxNgDbWV1CwTOBtFEVGA0DZile4vugOQNSOS9lSa10cDh5u0AKHjcS9bCGsgpts7B4w6RFAaAmHUOEceMN2-1vDFO46DVvmZFFzRuQYKlQr4Zajf3dyj-5z50GIqxDMZhG_u1p8r9eGEcZoparOzD-DEcV9w2K/s2560/2.jpg' },
-    { name: 'Filming Site, Suncheon', coords: [34.9089143, 127.34225201013444], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEg3uS8jNtUPWLAlS7MdMi2DS0ZW9KnYB__2LkndH74DCl7crvWOQrUoBpgbe-HMPOwnIDMj1xNASITK_G81g6o1cHzNXeDJkKrHX4fbPepMQAvmTx0kjKjVeSBcY7EsccwWy9-vmDKK8otxIcXqD7Xt4az-CGgs52QW-Sq0z3y300-KPzIISuRwt3Bq6D1C/s2560/photo_2024-10-30_00-16-51.jpg' },
-    { name: 'Eiffel Tower, Paris', coords: [48.8582599, 2.2945006358633115], imageLink: 'https://jasurlive.uz/img/pic2.png' },
-    { name: 'Vertical Forest in Milano City', coords: [45.45913914397633, 9.187069449510059], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEjopGWMQwXjFcqhYZwCIg2p0yTcxJap8yEQeZUBww_BUIDih3s0CirCYIOssac2OykY7Guoo0RaOUec7l3oT-6SvQkE8-Dm3YjrwEzZ14JMxku3Hnx3zm-QyBUbbbDpxKgtIyqOzISoG6XvzC1_ph-4a9U2QrRoFazPTL-OeF7Jl_AYAmh663JuQGsQ3gPA/w381-h508/bosco.jpg' },
-    { name: 'Galata Bridge', coords: [41.01978632234401, 28.97251904010773], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEjvCKEsz5vl8HZFdlUQiwQG6J_dbPXl22DdAhQ96W4-T1rAsREyUkXFAKvqTMNDtk4qR6KNqjY6fCOL7otlKgrbuiXXfCFNDL9WlInMLrUsOBExDtmJl4lOJORXpGWG3jb-jp5B2VI5oCRTMhhxcd-y1H9-J-xHap0xTPWu-mDYe-6p2WLkRTAD41q_DLLG/s320/20230927_152908.jpg' },
-    { name: 'Emirates Stadium', coords: [51.55504035, -0.1083996708672374], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEhOF6_gTwDvNkjcK6lfoibyExMVYRh1IlunmBwV0_qADTVTNYAh6-my7wVowz_vCl31owtAE9Cz2omvjsFEnOHNIK66UTl2d3NmezuYmZ5ZnaLgAtS85bqAX5l4RUB2ZQXivCQ1sGXQ7NLrr0kpE9wTGmvUGtD_mEBtZJGLDGMY7DMyV7ADRPGO9o_k7LKq/s320/23.webp' },
-    { name: 'Liverpool Lime Street', coords: [53.4076085, -2.9775854], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEjzfWXg9lQt2pj3wIbcyWnLnWPGU3qyxDVKw1Xgva7RI_wSc5RFMBCev-ODv5MkEsBztlv6M0u-OlgNl2Tw4TG_ZuqtDD4wbGRcz2ZUG2Uu7NgNBN8M77HLpAGjy2T2B9foJXdUS6tUpsr7L14egFpEDQLFNw-tc2-c9EI3VsaiERuW3Ej3LpI_vUEg46oY/s320/hachiko.png' },
-    { name: 'Stamford Bridge Stadium', coords: [51.481686499999995, -0.19103517666346065], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEghzbP3ppLImNNZQghA7R1TFTgPuPd0kIydru4EEDM7Tue8OOoi6KZhSzhQewXMyVAn6jXdXRqdrwxyKAi4EjCHdVzdIFOg6e9AXMopoSId2hyIBOh58kXeXc2sI7Vk_cu0lHJHmAtilhKYHOJgC79wvL4VSyLQPj_7b60VlOcz16LuRbp0vvfgnk4IdwSs/w437-h437/chel.jpg' },
-    { name: 'Santiago Bernabeu Stadium', coords: [40.4523667, -3.6907254], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEj8FzWKu5WDrW9ZYBBautwn7HRiuCUK6k-hI8UyM0jCi6cwo3lVim1U1ZuxWJlpszcQfX2tMRZvb4pMuqRKZDB2EIFJlf0AiEBzhoIo4gbB8MbcxFVbjIiEZe_Cdu69ALn-AgSXIbkrBwTZDrikLFm2PBrnGLqhg3uxxFogdyRpbz0nnjdNW4glAhWTuzc_/w384-h384/pic3.png' },
-    { name: 'Etihad Stadium 🩵', coords: [53.48309105, -2.2002520023433307], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEjiGGDH4O_MK5UeEVTn8viWKMtgf5PTLV_zbBxWRWsk7qzJmaTzmEYZ_D7wj_HBkWB0Ko_GchKRftbmj2nz5ySRPXgjb-ZfqemIOVRJ72tALF9iKp-T-YWKm6e_2C-WF2fScx7mlSyhE9nICV8QSQkwv8yw9uTAG_lEAi5YK1Ll9Gt3dK4WVMmiYZy2n-kY/s640/photo_2023-08-20_14-04-17.jpg' },
-    { name: 'Sherlock`s House, Baker Street', coords: [51.5237629, -0.1584743], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEgz50hezAN9SuvkYi9hQjLg2F-am3FSQiZBvxvPus-P8rcl0wSCZzsDVxJYGh2ppoh_coQJDCNMYMbM7f8wwS6tlxQ1Pjm28ucYmcOTEqk-q4a8O6IkVKMDjncffzg_-lQqXzxB7-4vY0K3pKUoh0HlV9GxBr4pY0M3vLTTiiFr-nyWRWLDYxogafdGLiow/w277-h616/20221222_113534.jpg' },
-    { name: 'El Principe Restaurante, Madrid', coords: [40.4141634, -3.7000816], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEizSdkVDVxnrCMhs6bMX2so39R4-U_oEWCVd6D4SzKjNCEPGH4GD0hI14HRtaQK-Gnb48cY70aLXj49eeoGNSyOD9YG5jM7gepQW_E3Rmvrrxhbw9bbmQKe68PSzXlZbqOEVyJj3puBcBcJOdRYYfu4jbVVCHvv4OsaHrjNizzLbIu_Q2i-4tUlftk1CQrf/w497-h663/20230811_220557.jpg' },
-    { name: 'Wanda Metropolitano Stadium', coords: [40.43503136007304, -3.600726127624512], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEjDsFcxaZ6TphewzLHXroTjHNJrp3ha3ks8Ky8kI5wg2c_yy1e9rpemwSfs1bcE2P5yvXtuhq1gB2ZNEPAkMQnBI_M8440xCs8wtAcZK0Es_Jkf7lfko1lI3g1TtjwUF_HFjtuwj4SYgAz7yFh4PBOoYR2rwFJ-GhzmOPUqnNyXIxJCGlhw1BgbsmwAZBvO/s320/photo_2023-08-24_17-47-13.jpg' },
-    { name: 'University of Liverpool ❤️🎓', coords: [53.40611160109354, -2.966246349630762], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEgZ-gSKxKu9nN59x3CDw61foCAWR227Q_MrNAS2RRJ5WyjmGd4hQUaT6w8fyiM608ksQZ32r0l9WZtwE2h03T2Ep7hGMXzi8KIC_Gefk69gpBSEIOrqxwmBMREhsgEmO6IK0LawPDSHG_TQDyAmwK5ROxKtDkvG1NdWmZIF_mjqpuplzz5YFGElx-fDrDM7/s320/121.webp' },
-    { name: 'Beautiful Chester 😻', coords: [53.1908873, -2.8908955], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEhXOrbZd07ixubpijDX8fo5Zum4rE3dSNpvVca4Rup-TmW3ggdrZOZFYRKBj6oFMQjb3WGhg5KSU4R_M1aLS7kxlbdNYTWH5cxcuY7ZboxK5aQBeLQF-25T1GdxMViGi7L1WtE9buHBELopevjg-k8eIFduZDybUrHhhN2qn6CATW9fz4kd9OmbFbpsjL_4/w440-h587/photo_2024-07-28_15-33-00.jpg' },
-    { name: 'George Square, Glasgow', coords: [55.86115665, -4.250218902542482], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEg5yiegpbL8XJfIfj2nRGM3s-18Ysc4ipJyUG5SJpabX_h7hyphenhyphenF9InYU22iUplermNTwbbyKBIkqZQgdLS9HmGbcTsshoV4s5iwVQnxckb0klBUOBdkm_y-NqoQVN0bFX8oJZ_SInll5lt-bLxQ3ykgcVBj1iWubO5B2IOShqD1nF1vNDq6K1i7qjv3LS-rB/w435-h580/photo_2024-07-28_15-32-50.jpg' },
-    { name: 'San Siro Stadium 🏟️', coords: [45.478200349999995, 9.123963993342393], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEgASsUAt2k9s1V8t78KAMkhpLM-bth5LCaFp5v7oFsUWkc4KnujDWWS7qc8UaqMSl6s5OJ39DxwaCHh97cafsbHJRqaNFGUcNwBZdgQoAjdsvhcDIbbWAtAxYpm7LXRKJhrio_q4yK2lOnM-ix4NMu2jUsdG7pFS_52QRS-6q3fS-IhUco8TRESglM-0TBw/w414-h449/photo_2024-07-28_15-32-38.jpg' },
-    { name: 'Llandudno, Wales ', coords: [53.33151996613546, -3.8251411914825444], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEj6V33tqCggT0H5B5zGrXk4OuugyLQNIF-ROKt9_DtDsn4WSIEvhD6zzVdS0p1rwx4cI857c9ZXv6HB9zzvZVA7Gf0rkXsnNKt288HXI-B1_B577lGsgaKwruKP0C0PjvEwFpeeaajOA5mq0jRgVQsSPKQhRryBMVgRdEisQjSGZVC69pfM8LL9WV6PxEol/w438-h584/photo_2024-07-28_15-32-47.jpg' },
-    { name: 'Edinburgh Castle ', coords: [55.948686699999996, -3.2004174629518065], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEhCOLr1lTkfIzzJbbIvUcfumVGZ-mLlyUny5P3-pKCaCW7kTeGoFfs1KGJ939XmcaSRty1r_wmhnNiurcd7BCcJI19R_FgEDM_rbrMUksCo9kfpuxUIIsFsVpoaXp0Z1bnOolreBMTSv9EN1SPxmB91jXUyFGoz1W1U-PDWQxA2S7srZwAhs1foWYRb9Jdt/w459-h612/photo_2024-07-28_15-32-55.jpg' },
-    { name: 'Samarkand Railway Station', coords: [39.6855379, 66.9289120568356], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEjb6ZSz8-QnKLB751LR7hRAaY62805fwE6tTUUUHIfgjtbeomQ-hKK9ReekLHtuya0-mdrW4rBU3_OrCXEhlh7kNZESDO1JLzsGlbCoF65yYeg73ERFME05aD4i1wadNYOQ3_y6x1SngGjsKdA-UHWr4wkuyrpmZFM5x-Fo0vLgWVKFEU2V1zrLYoqadR9a/w428-h570/photo_2024-07-28_15-32-33.jpg' },
-    { name: 'Old Trafford Stadium 🏟️', coords: [53.46310955, -2.2913864850545362], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEhofbIsGXjPi-qObA45gHGGkZFkmf9zLcmsllapOi8MsSbA2mLZffGD_1EgAAbN5ajXOlzANyJrL8SYV_Q-JcTXzBt7zZflGiYffvUmzhRXEkW03apyIObLzOO7oPXho18o2GN1XlvOBuy4_wMZ3WGOC9Ao3x9X_g1cyuEUR79sFpJyo-Ci8cICo1YbM2oE/w448-h597/photo_2024-07-28_15-32-42.jpg' },
-    { name: 'Nyhavn, Copenhagen ❤️', coords: [55.67973595, 12.590885478619983], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEioRIb20PjXB1oG13XS79VXJffXelyRTEp6snx4fRLH9ttDH9irzkqS14plHQbPmeoc3phItgPCT6CsoeP5WHRTyedOssBcyokGkgToorKrL6I1CMbaezGPg3vBzBw_ahNu_kee3DpGpOQqdv-2Ok8KGZ8IPIiKB-SQ4vYlQf6TeR_NzTuY84Mma33z067j/w601-h411/photo_2024-07-28_15-49-17.jpg' },
-    { name: 'Sleep In Heaven 📸🏆', coords: [55.6867883, 12.5505333], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEg4_aWNSB-Qt1bI7on2JRSvKoBXIXAVVnTEtRfUAFmppKhXM0NGYwVRqvPuIoIUMQyX9KJIYGUyZhBmChTVaCHtkkOimkzh8dHy7he0ksmYljn6Cx4xCBD30zfh7_rQ8pO_CyzZqmaG5y0BAhHiLbmgjITUaFaVPbqlTFBg_2C1Kb3G3Sll3jfdXpAfjhZj/w519-h692/photo_2024-07-28_15-51-54.jpg' },
-    { name: 'Copenhagen Airport', coords: [55.62565886710453, 12.648696899414064], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEhNc95qezJx9QQLvu54uo7g7Mht-rhMfQfzqPAcPALpeQ_lUWha33-PVJTOvUmFGxcUbxHGSXIDW2DET7I3oR6dX5rj4e8ThFcOMshrFkTg7n4HGpuGFYuEgYDKljJZWwRM1mmjze85WO5eSSpJ0Zi9seKapCghFYxKVKtzFlT9uIVAGbOJDq4WtodUjxEA/w458-h682/photo_2024-07-28_15-55-59.jpg' },
-    { name: 'Tashkent State Transport University 🎓', coords: [41.27729075, 69.2826682070976], imageLink: 'https://tstu.uz/assets/front/images/tstuphoto.jpg' },
-    { name: 'Tashkent Railway Station', coords: [41.2932886, 69.2877212], imageLink: 'https://upload.wikimedia.org/wikipedia/commons/4/45/Tashkent_Railway_Station.jpg' },
-    { name: 'Uzbekistan Locomotive Depot', coords: [41.17858174722078, 69.12394344806673], imageLink: 'https://tashtrans.uz/photo/railway/depo/DSC_0004.jpg' },
-    { name: 'Bulungur 💚', coords: [39.76059340816414, 67.27417945861818], imageLink: 'https://i.ytimg.com/vi/5ko2xpNPV4U/maxresdefault.jpg' },
-    { name: 'SKD Airport Samarkand', coords: [39.69887605, 66.99122805921263], imageLink: 'https://upload.wikimedia.org/wikipedia/commons/thumb/2/2e/Samarkand_airport_%28SKD_-_UTSS%29_Entrance_from_airfield_31.05.2022.jpg/640px-Samarkand_airport_%28SKD_-_UTSS%29_Entrance_from_airfield_31.05.2022.jpg' },
-    { name: 'TAS Tashkent International Airport', coords: [41.2599624, 69.28511704456324], imageLink: 'https://people-travels.com/storage/images/sights/design/Tashkent_airport_1.webp' },
-    { name: 'Daejeon, South Korea', coords: [36.3352732, 127.45293497231432], imageLink: 'https://facts.net/wp-content/uploads/2023/07/50-facts-about-daejeon-taejon-1688365652.jpg' },
-    { name: 'Almaty Airport', coords: [43.343207050000004, 77.01258085779281], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEgjcZqq-VLge0IYDxFOYHwHEZQVn68YJ3kWtwpw5T77P3zd_fvPb21nuo6TgiuEjYQUsKmnaCjwim-b2oX7fvVq-yW1cHgxDXRUXb7xhsRJ1uHBvBtevXLfbY8ox7CzT02psx50fsfTaoq-MuFT-EtyVvT7MWZtxyY8CHjbgfMph4khM1Dt8xCW1k8sGopg/w445-h782/photo_2024-09-13_16-16-42.jpg' },
-    { name: 'Cafe in Gyeonggi-do | Kimchi time 😁🌶️🍜', coords: [37.1994653, 126.8312632], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEjG6n8cDzbBJlYgXy5JyCeHzK9RVf2KY4Xunm-KueVtDnZrr0mrlSo1LKihpK2ckRSgrqWN-vC6hY8bmcNdBe5zqpdvTduwNywBUphdPJNXGjhc6-hWDnYjnEgqaa4OvocBhbCBIfl1IsB1xGTLs4OquKdisMteME05Nu5DuCVq-4dj3DoV9oHeBF056ui_/w437-h544/photo_2024-09-13_16-15-48.jpg' },
-    { name: 'Woosong University West Campus', coords: [36.33772765, 127.44596583060166], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEhmq7F6Btcp6ilAMW2Wz89XVrvKmXiXIPfBlplUie2olI49DyxUIaUIBY5UA3k0ckiAXTA2gA5SyeLWahNaV_Dx951_9iaE9PhPcRWoMU1qC4aY_njfyRZQz2YnNUxKaoBEjpbeClMzAPZ1H4nmMdx4PJEMXYbwGwF3KDYRIl6UaKraBigH6apkIm8oRqC1/w442-h439/photo_2024-09-16_13-43-34.jpg' },
-    { name: 'Uam History Park', coords: [36.356665642349476, 127.43720591068269], imageLink: 'https://blogger.googleusercontent.com/img/a/AVvXsEhgMOFWMnHGmU9s4mUIWmsU8hMS1KMbnjGfDRfzHesRPOn7m1SpUSaVvn2KZB8E15PZ77e0e810OMY6mo_fgdZANmio-9YotxlPI_J9NrSkrRygZKikcRo1emjnR7dL50W43ihgY-HMAM9Q7VeTTOUuRscyZhpLX3bkHaL4nxJ8JVIxmo7ATBJbZNW6IJl0=w387-h555' },
-    { name: 'Solbridge, Woosong University', coords: [36.338327750000005, 127.43243066372395], imageLink: 'https://blogger.googleusercontent.com/img/a/AVvXsEgfp-AToYuhE_mDHxcvhi1aKbcWHCbupZ8A5AgejcVi4HKKV_dzH0-kqnv8c31GaAHz-Dr3tJRhJlPht2XD0NcKxqhEVfOyvMLb5DHosI-Sw3bNhqnpBahJbF9EOwdFqRShMco90wJ525ferP_WTtsMEgWewOrWjzIR6g3yNHBWO-GIqYZE042qTs5VHLKA=w382-h488' },
-    { name: 'Daejeon Expo Science Park', coords: [36.37600915, 127.38768740014706], imageLink: 'https://blogger.googleusercontent.com/img/a/AVvXsEg6b2BV71DK3MlG8zyCs8E4v4ste-IQ4m4NbQwggsUYQAQ9TeEyy4kpspT_DqBa3H9J5s_yja7KrZYGT-X4VL79MVSs-BVzsDw-BodwTICzgMhTv-D9vtH3rcGrtwgPNdXygOme6JDBLnMNNsh1dgHyjFq2amCHb9RoXBan7Wnpp_cS4euLSz84iXr8ix4X=w318-h479' },
-    { name: 'Daecheongho Lake', coords: [36.41819665, 127.49738476559037], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEjbHbKIBCP1qa8IfejkWYGE2__7qMxJekq07IbedqZkPgmnTWu1fWq-46OMstXV9bbFvJIj5fIfxeqvr46mi5eWby1bsOFPLy_Gn5eDyfbHKCmjYQeCDPo4uSNswnxrW7r2J9YG0_8dYDsSfNbZB8hKxVTgVK7gDO5yB8psBAELWhb_onAP_HrpCySa0nfz/w466-h466/dw.jpg' },
-    { name: 'Yuseong Foot Bath', coords: [36.3549475, 127.3460614], imageLink: 'https://blogger.googleusercontent.com/img/a/AVvXsEiJFEbtWAeRDeWKubd4tfHBk9DOQ3x92XWYvZ4U_iQtC30dKlY4cG94P-OULZ8EeQaVzdUYcl3yw_U9lix58tYIOqY7SCS88Fi0vewkC2UeVZWqhg7i9S4FcHIFVc6G9fUdaESYQZ2gZL3CqssO9myYmeb0lFDwYX5vuMk7VtARJ0Kl-J06GXIle5rQDXN0=w388-h388' },
-    { name: 'Daejeon (Oncheon 1(il)-dong), South Korea 🇰🇷', coords: [36.35591446332886, 127.34676445038767], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEguqclW9vE_gcwR0rEEMKRpSsYC51FA6hkeKv0g0iQWkra2rtKSsHZLrqMvoQBi8gglifne6XJhO3Y5KxZbpS-fXOVcLdtg4OupmcETgFmJjpZwnUJE82QB6wflRWyPrvLPlvJTnzSX5xNyAlTVbjg_LWalPHqD-OXvnmt9sj7sikxFg70JU3P3wnWkmTZZ/w261-h440/20240927_182946.jpg' },
-    { name: 'W7 Woosong University', coords: [36.33705931024952, 127.44506729398547], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEgxPjgh8tW56xU7hBA4Or6EhGhGTvUou_H1RqYat2x3pH0h-tajLow5vgbXH3F274U4rWjcQ3W-oS6Xlw0Nfc3lq2P8x3fEmgwdIolL-leqeQHyX2iZL_VDlk0le8l6shDp5ZpCXpq8EWWykpARwjUnMIkvwnlzIa-vRDmLqreR5Qeg9nOIwYtH8HnfA2mD/s1280/photo_2024-10-04_02-08-52.jpg' },
-    { name: 'Daedong Hanuel Park', coords: [36.33296217972606, 127.45091472323428], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEhd-jFLheKE7iYUKNc2jtfZGY9A1mVr1BVI1575CkRUQtp2i11jt1RReSkD_tGL_-zI31B-_cLc8yriH52JrpS4HHdCNDrpnY7qFe-ToE-jbwDCDQnKTWXCyMGPQGSr_GZPlfzJs4uBSQLjNUDLmvH7SKINLW7iuajVFbtgozL1sqrarz1sOtyZWn3EvtoI/s1527/autumn.jpg' },
-    { name: 'Starfield, Suwon', coords: [37.2874448, 126.99157523573354], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEiCABkUyVZ0_PpAuBXPAuvkyWRfR2bYF6beUpN0gJqVzFc0zwtMuGACl8DkuHk-REb39MrGmqGnQBVu7CoACwJF40ww-4ZJja44RzOxkuXaPlK-cDFLzFENGoP2TDe_nC7bDmL_vnIEmj4WiuCUb9piRXJX2u7FVupS1ep_szSjmEtMnfCxDCxwUUmRrJQa/s1426/camera-print%20(4).png' },
-    { name: 'Ppuri Park, Daejeon', coords: [36.284136700000005, 127.38785219063291], imageLink: 'https://blogger.googleusercontent.com/img/b/R29vZ2xl/AVvXsEiAB0Jd1Z6hBaCohTTbgrMFoY-rm-5wqGdvcqnn_ZFJw-c9FaHHL3unH0qIgpzWFBc3rFQbdxat8lvVLRbtki528lrXHp6ewFphaU1W75x_ta8NE8uhksuiyLsz2zxq23RIkGK9N2wkjPzxPnVozWiyCqP20ey4zVhCKWBf0TWN55Nc-OLHWH1094fBWGzo/s1782/camera-print%20(2).png' },
+      // Fetch the data from the "Planned" sheet
+      const plannedSheetName = "Planned";
+      const plannedSheet = workbook.Sheets[plannedSheetName];
+      const plannedData = XLSX.utils.sheet_to_json(plannedSheet);
 
+      const visited = [];
+      const planned = [];
 
+      // Process the visited data
+      visitedData.forEach((row) => {
+        const coords = row.Coords ? JSON.parse(row.Coords) : [0, 0];
+        const placeData = {
+          name: row.Name,
+          coords: coords,
+          imageLink: row["Image Links"],
+        };
+        visited.push(placeData);
+      });
 
-  ];
+      // Process the planned data
+      plannedData.forEach((row) => {
+        const coords = row.Coords ? JSON.parse(row.Coords) : [0, 0];
+        const placeData = {
+          name: row.Name,
+          coords: coords,
+          imageLink: row["Image Links"],
+        };
+        planned.push(placeData);
+      });
 
-  // ------------------------------------------------------------------------------------------------------- ✈️ PLANNED PLACES BELOW 
+      // Set the state for visited and planned places
+      setVisitedPlaces(visited);
+      setPlannedPlaces(planned);
 
-  const plannedPlaces = [
-    { name: 'Tokyo, Japan', coords: [35.6821936, 139.762221], imageLink: 'https://www.travelandleisure.com/thmb/VscfMXHNO6uBpaX2cuIW1q2ZtA8=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/tokyo-japan-TOKYOTG0621-52012ff551dc46c4a87ac8e3151307a4.jpg' },
-    { name: 'Phuket, Thailand', coords: [7.879982497172296, 98.38960647583008], imageLink: 'https://miro.medium.com/v2/resize:fit:1000/0*fY1_mh4P45v1EYdX.jpg' },
-    { name: 'Kuala Lumpur, Malaysia', coords: [3.1516964, 101.6942371], imageLink: 'https://media.timeout.com/images/105866262/750/562/image.jpg' },
-    { name: 'Sydney, Australia', coords: [-33.8698439, 151.2082848], imageLink: 'https://static1.squarespace.com/static/55ee34aae4b0bf70212ada4c/57d9829837c5819632bc630b/5e06e89e74877f00dbb21494/1577545446402/keith-zhu-qaNcz43MeY8-unsplash+%281%29.jpg?format=1500w' },
-    { name: 'Mecca, Saudi Arabia', coords: [21.420847, 39.826869], imageLink: 'https://cdn.britannica.com/20/153520-050-177A78E4/Kabah-hajj-pilgrims-Saudi-Arabia-Mecca.jpg' },
-    { name: 'New York, Liberty Statue', coords: [40.689253199999996, -74.04454817144321], imageLink: 'https://www.worldatlas.com/upload/f4/d8/7b/shutterstock-1397031029.jpg' },
-    { name: 'Hawai 🌴🏖️🍹', coords: [19.593801499999998, -155.42837009716908], imageLink: 'https://www.royalhawaiianmovers.com/wp-content/uploads/2021/11/111521-Best-Island-to-Live-on.png' },
-    { name: 'Buenos Aires, Argentina', coords: [-34.609023, -58.3812732], imageLink: 'https://i.natgeofe.com/n/a2e54fd2-f71b-420b-b9dd-8a35cc5c6f75/ArgentinaLead_2x1.jpg' },
-    { name: 'Hawai', coords: [19.588489212685314, 204.58153972248294], imageLink: 'https://www.royalhawaiianmovers.com/wp-content/uploads/2021/11/111521-Best-Island-to-Live-on.png' },
-    { name: 'The Great Pyramid of Giza', coords: [29.97916465, 31.13421326082681], imageLink: 'https://media.architecturaldigest.com/photos/58e2a407c0e88d1a6a20066b/16:9/w_1280,c_limit/Pyramid%20of%20Giza%201.jpg' },
-    { name: 'Medina, Saudi Arabia', coords: [24.471153, 39.6111216], imageLink: 'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/13/60/b1/42/medina.jpg' },
-    { name: 'South Pole Point', coords: [-66.28388122840676, 53.69731440027301], imageLink: 'https://waterproof-expeditions.com/wp-content/uploads/2024/02/South-Pole-Antarctica-5.jpg' },
-    { name: 'Cape Town, South Africa', coords: [-33.928992, 18.417396], imageLink: 'https://static.independent.co.uk/s3fs-public/thumbnails/image/2019/04/15/13/istock-671323338.jpg' },
+    } catch (error) {
+      console.error("Error fetching or parsing Excel file:", error);
+    }
+  };
 
-
-
-
-  ];
+  useEffect(() => {
+    fetchExcelData(); // Fetch and process data once the component is mounted
+  }, []);
 
   return (
     <div>
       <h1>My Travel Map 🗺️ == 🚩-visited | ✈️-planned</h1>
+      {/* Render Map component with fetched data */}
       <Map visitedPlaces={visitedPlaces} plannedPlaces={plannedPlaces} />
     </div>
   );
 };
-
 
 export default HomePage;
